requirements.txt
# Existing dependencies
apscheduler
pytz
tzlocal
requests
flask
path

# New dependencies for platform upgrade
# Database
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
sqlalchemy>=1.4.0
alembic>=1.8.0

# Cache
redis>=4.5.0

# Async Task Queue
celery>=5.3.0
celery[redis]>=5.3.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data validation
pydantic>=1.10.0
pydantic[email]>=1.10.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Async support
aiofiles>=23.0.0
httpx>=0.24.0

# Monitoring
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk>=1.40.0

# Analytics and Reporting
reportlab>=4.0.0

# Machine Learning (Cookie Classification)
scikit-learn>=1.5.0
xgboost>=2.0.3
pandas>=2.2.0
numpy>=1.26.3
joblib>=1.3.2
shap>=0.44.0
matplotlib>=3.8.2
seaborn>=0.13.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0